Subdivide `index.mts`
