+++ src/events/discordjs/guildMemberAdd.mts
@@ -1,0 +1,10 @@
+import {Client, Events} from "discord.js";
+import {flags, getFlag, setFlag} from "#src/agents/flagAgent.mjs";
+
+export function init(client: Client): void {
+    client.on(Events.GuildMemberAdd, async (member) => {
+        // Make sure the user isn't considered a ghost
+        let wasGhost = await getFlag(member.id, flags.Ghost);
+        if (wasGhost) await setFlag(member.id, flags.Ghost, "false");
+    });
+}
